begin implementing a cli
This is mostly a thought experiment, and a chance to write some Rust.
As a first draft I thought a reasonable endpoint might be the two
opposite points of a rectangle within which the probe will take a
set of measurements.  An obvious extension would be to specify
resolution.
For the moment this provides a framework to think about the interface
and a small project to write some Rust in.